requirements.txt
boto3
h2o==3.24.0.2
matplotlib
numpy==1.15.4
pandas
pyyaml==3.13
requests==2.20.0
sagemaker
seaborn
scikit-learn>=0.19
tables>=3.5.1
